doing cheaper, probabilistic checks outside the scattergl conversion loop
